Import ACPICA 20130626.

+++ source/include/acutils.h
@@ -903,7 +903,7 @@ AcpiUtStrtoul64 (
 void
 AcpiUtPrintString (
     char                    *String,
-    UINT8                   MaxLength);
+    UINT16                  MaxLength);
 
 void
 UtConvertBackslashes (
